requirements.txt
bcrypt==4.2.1
anthropic>=0.83.0
boto3==1.36.22
botocore==1.36.22
build==1.2.2.post1
click==8.1.8
cryptography==44.0.3
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==6.0.0
Flask-Injector==0.15.0
Flask-Session==0.8.0
flatbuffers==24.3.25
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth>=2.46.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-genai==1.57.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx>=0.28.1,<1.0.0
httpx-sse==0.4.0
huggingface-hub>=0.16.4
humanfriendly==10.0
idna==3.10
injector==0.22.0
Jinja2==3.1.5
langchain==0.3.19
langchain-core==0.3.35
langchain-text-splitters==0.3.6
markdown2==2.5.3
openai==2.8.1
openpyxl==3.1.5
pandas==2.3.1
pgvector==0.3.6
pillow==11.0.0
psutil==7.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMuPDF==1.25.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytesseract==0.3.13
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.11.2
sib-api-v3-sdk==7.6.0
SQLAlchemy==2.0.36
tiktoken==0.8.0
tokenizers>=0.22.0
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
pyjwt[crypto]>=2.8.0
